Tidy up metric API / add interface check
