+++ api/metric/noop_meter.go
@@ -6,7 +6,13 @@ import (
 	"github.com/open-telemetry/opentelemetry-go/api/core"
 )
 
+type noopMeter struct{}
+
 type noopMetric struct{}
+
+var _ Meter = noopMeter{}
+
+var _ Float64Gauge = noopMetric{}
 
 func (noopMeter) GetFloat64Gauge(ctx context.Context, gauge *Float64GaugeHandle, labels ...core.KeyValue) Float64Gauge {
 	return noopMetric{}
